moved collections model into a controller
