--- src/Submissions/Api/Areas/Biobank/Models/AddSampleSetModel.cs
@@ -1,5 +1,0 @@
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
-
-public class AddSampleSetModel : AbstractCRUDSampleSetModel
-{
-}
