+++ src/Submissions/Api/Areas/Biobank/Models/Collections/AbstractCRUDCollectionModel.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
 
 public class AbstractCRUDCollectionModel : AbstractCRUDWithAssociatedDataModel
 {
