+++ src/Submissions/Api/Areas/Biobank/Models/Collections/AddCapabilityModel.cs
@@ -1,6 +1,6 @@
 using Biobanks.Submissions.Api.Models.Shared;
 
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
 
 public class AddCapabilityModel : AbstractCRUDCapabilityModel
 {
