+++ src/Submissions/Api/Areas/Biobank/Models/Collections/AddCollectionModel.cs
@@ -1,0 +1,5 @@
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
+
+public class AddCollectionModel : AbstractCRUDCollectionModel
+{
+}
