+++ src/Submissions/Api/Areas/Biobank/Models/Collections/AddSampleSetModel.cs
@@ -1,0 +1,5 @@
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
+
+public class AddSampleSetModel : AbstractCRUDSampleSetModel
+{
+}
