+++ src/Submissions/Api/Areas/Biobank/Models/Collections/CollectionModel.cs
@@ -2,7 +2,7 @@ using Biobanks.Submissions.Api.Models.Biobank;
 using System;
 using System.Collections.Generic;
 
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
 
 public class CollectionModel
 {
