+++ src/Submissions/Api/Areas/Biobank/Models/Collections/CopySampleSetModel.cs
@@ -1,4 +1,4 @@
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
 
 public class CopySampleSetModel : AbstractCRUDSampleSetModel
 {
