+++ src/Submissions/Api/Areas/Biobank/Models/Collections/SampleSetModel.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Biobanks.Submissions.Api.Areas.Biobank.Models;
+namespace Biobanks.Submissions.Api.Areas.Biobank.Models.Collections;
 
 public class SampleSetModel
 {
