+++ src/Submissions/Api/Areas/Biobank/Views/AddCollection.cshtml
@@ -1,4 +1,4 @@
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.AddCollectionModel
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.AddCollectionModel
 @{
     ViewBag.Title = "Add collection";
 }
