+++ src/Submissions/Api/Areas/Biobank/Views/Collection.cshtml
@@ -1,7 +1,7 @@
 @using Biobanks.Submissions.Api.Config
 @inject Biobanks.Submissions.Api.Services.Directory.IConfigService _config
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.CollectionModel
 
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.CollectionModel
 @{
 	ViewBag.Title = "Collection";
 }
