+++ src/Submissions/Api/Areas/Biobank/Views/EditCollection.cshtml
@@ -1,4 +1,4 @@
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.EditCollectionModel
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.EditCollectionModel
 @{
     ViewBag.Title = "Edit collection";
 }
