+++ src/Submissions/Api/Areas/Biobank/Views/Index.cshtml
@@ -1,5 +1,5 @@
 @using System.Linq
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.BiobankCollectionsModel
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.BiobankCollectionsModel
 @{
 	ViewBag.Title = "Collections";
 }
