+++ src/Submissions/Api/Areas/Biobank/Views/SampleSet.cshtml
@@ -1,7 +1,7 @@
 @inject Biobanks.Submissions.Api.Services.Directory.IConfigService _config
 @using Biobanks.Submissions.Api.Config
 @using Biobanks.Submissions.Api.Constants
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.SampleSetModel
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.SampleSetModel
 @{
     ViewBag.Title = "Sample set";
 }
