+++ src/Submissions/Api/Areas/Biobank/Views/_CollectionSharedFields.cshtml
@@ -1,5 +1,5 @@
 @using System.Configuration
-@model Biobanks.Submissions.Api.Areas.Biobank.Models.AbstractCRUDCollectionModel
+@model Biobanks.Submissions.Api.Areas.Biobank.Models.Collections.AbstractCRUDCollectionModel
 
 <div id="bloodhound" class="form-group">
     <label for="Diagnosis" class="col-sm-4 control-label required">
